+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Support `:scope` selector
 * Bug fixes to ~= operator, which now accepts non-space whitespace as a
   separator and is stricter about match targets containing whitespace
+* Rudimentary namespace selector support: `*|TAG` and `|TAG`.
 
 # zest-css 1.1.5 (2021-03-23)
 * Dependency updates